+++ plugins/processing/signal-processing/src/box-algorithms/basic/ovpCBoxAlgorithmIdentity.cpp
@@ -1,7 +1,7 @@
 #include "ovpCBoxAlgorithmIdentity.h"
 
 using namespace OpenViBE;
-using namespace Kernel;
+using namespace /*OpenViBE::*/Kernel;
 using namespace Plugins;
 using namespace OpenViBEPlugins;
 using namespace SignalProcessing;
@@ -19,15 +19,15 @@ bool CBoxAlgorithmIdentity::processInput(const size_t /*index*/)
 bool CBoxAlgorithmIdentity::process()
 {
 	IBoxIO* boxContext    = getBoxAlgorithmContext()->getDynamicBoxContext();
-	const uint32_t nInput = getBoxAlgorithmContext()->getStaticBoxContext()->getInputCount();
+	const size_t nInput   = getBoxAlgorithmContext()->getStaticBoxContext()->getInputCount();
 	uint64_t tStart       = 0;
 	uint64_t tEnd         = 0;
-	uint64_t size         = 0;
+	size_t size           = 0;
 	const uint8_t* buffer = nullptr;
 
-	for (uint32_t i = 0; i < nInput; ++i)
+	for (size_t i = 0; i < nInput; ++i)
 	{
-		for (uint32_t j = 0; j < boxContext->getInputChunkCount(i); ++j)
+		for (size_t j = 0; j < boxContext->getInputChunkCount(i); ++j)
 		{
 			boxContext->getInputChunk(i, j, tStart, tEnd, size, buffer);
 			boxContext->appendOutputChunkData(i, buffer, size);
