+++ plugins/processing/streaming/src/box-algorithms/ovpCBoxAlgorithmStreamedMatrixMultiplexer.cpp
@@ -3,7 +3,7 @@
 #include <system/ovCMemory.h>
 
 using namespace OpenViBE;
-using namespace Kernel;
+using namespace /*OpenViBE::*/Kernel;
 using namespace Plugins;
 
 using namespace OpenViBEPlugins;
@@ -13,7 +13,7 @@ bool CBoxAlgorithmStreamedMatrixMultiplexer::initialize()
 {
 	m_lastStartTime = 0;
 	m_lastEndTime   = 0;
-	m_headerSent   = false;
+	m_headerSent    = false;
 
 	return true;
 }
@@ -28,12 +28,12 @@ bool CBoxAlgorithmStreamedMatrixMultiplexer::processInput(const size_t /*index*/
 
 bool CBoxAlgorithmStreamedMatrixMultiplexer::process()
 {
-	IBoxIO& boxContext    = this->getDynamicBoxContext();
-	const uint32_t nInput = this->getStaticBoxContext().getInputCount();
+	IBoxIO& boxContext  = this->getDynamicBoxContext();
+	const size_t nInput = this->getStaticBoxContext().getInputCount();
 
-	for (uint32_t i = 0; i < nInput; ++i)
+	for (size_t i = 0; i < nInput; ++i)
 	{
-		for (uint32_t j = 0; j < boxContext.getInputChunkCount(i); ++j)
+		for (size_t j = 0; j < boxContext.getInputChunkCount(i); ++j)
 		{
 			const IMemoryBuffer* iBuffer = boxContext.getInputChunk(i, j);
 			const uint64_t tStart        = boxContext.getInputChunkStartTime(i, j);
